Move utlity function near its users
It is not used by the authorizer anymore.
