+++ pkg/authorization/registry/localsubjectaccessreview/rest_test.go
@@ -12,8 +12,8 @@ import (
 	apirequest "k8s.io/apiserver/pkg/endpoints/request"
 
 	authorizationapi "github.com/openshift/origin/pkg/authorization/apis/authorization"
-	"github.com/openshift/origin/pkg/authorization/authorizer"
 	"github.com/openshift/origin/pkg/authorization/registry/subjectaccessreview"
+	"github.com/openshift/origin/pkg/authorization/registry/util"
 )
 
 type subjectAccessTest struct {
@@ -295,7 +295,7 @@ func (r *subjectAccessTest) runTest(t *testing.T) {
 		t.Errorf("Unexpected obj type: %v", obj)
 	}
 
-	expectedAttributes := authorizer.ToDefaultAuthorizationAttributes(r.expectedUserInfo, r.reviewRequest.Action.Namespace, r.reviewRequest.Action)
+	expectedAttributes := util.ToDefaultAuthorizationAttributes(r.expectedUserInfo, r.reviewRequest.Action.Namespace, r.reviewRequest.Action)
 	if !reflect.DeepEqual(expectedAttributes, r.authorizer.actualAttributes) {
 		t.Errorf("diff %v", diff.ObjectGoPrintDiff(expectedAttributes, r.authorizer.actualAttributes))
 	}
