+++ pkg/authorization/registry/resourceaccessreview/rest.go
@@ -14,6 +14,7 @@ import (
 	authorizationapi "github.com/openshift/origin/pkg/authorization/apis/authorization"
 	authorizationvalidation "github.com/openshift/origin/pkg/authorization/apis/authorization/validation"
 	"github.com/openshift/origin/pkg/authorization/authorizer"
+	"github.com/openshift/origin/pkg/authorization/registry/util"
 )
 
 // REST implements the RESTStorage interface in terms of an Registry.
@@ -61,7 +62,7 @@ func (r *REST) Create(ctx apirequest.Context, obj runtime.Object, _ bool) (runti
 		return nil, err
 	}
 
-	attributes := authorizer.ToDefaultAuthorizationAttributes(nil, resourceAccessReview.Action.Namespace, resourceAccessReview.Action)
+	attributes := util.ToDefaultAuthorizationAttributes(nil, resourceAccessReview.Action.Namespace, resourceAccessReview.Action)
 	users, groups, err := r.subjectLocator.GetAllowedSubjects(attributes)
 
 	response := &authorizationapi.ResourceAccessReviewResponse{
