+++ pkg/authorization/registry/subjectaccessreview/rest_test.go
@@ -13,7 +13,7 @@ import (
 	kapi "k8s.io/kubernetes/pkg/api"
 
 	authorizationapi "github.com/openshift/origin/pkg/authorization/apis/authorization"
-	"github.com/openshift/origin/pkg/authorization/authorizer"
+	"github.com/openshift/origin/pkg/authorization/registry/util"
 )
 
 type subjectAccessTest struct {
@@ -271,7 +271,7 @@ func (r *subjectAccessTest) runTest(t *testing.T) {
 		t.Errorf("Unexpected obj type: %v", obj)
 	}
 
-	expectedAttributes := authorizer.ToDefaultAuthorizationAttributes(r.expectedUserInfo, kapi.NamespaceAll, r.reviewRequest.Action)
+	expectedAttributes := util.ToDefaultAuthorizationAttributes(r.expectedUserInfo, kapi.NamespaceAll, r.reviewRequest.Action)
 	if !reflect.DeepEqual(expectedAttributes, r.authorizer.actualAttributes) {
 		t.Errorf("diff %v", diff.ObjectGoPrintDiff(expectedAttributes, r.authorizer.actualAttributes))
 	}
