+++ pkg/authorization/registry/util/attributes.go
@@ -1,4 +1,4 @@
-package authorizer
+package util
 
 import (
 	"strings"
